refactor: separate pygame helpers
