+++ gomoku/scripts/play_with_pygame.py
@@ -9,73 +9,17 @@
 import pygame
 import sys
 import time
-import numpy as np
 from pathlib import Path
 
 # 新しいパッケージ構成に合わせてインポート
 from ..core.gomoku_env import GomokuEnv
-from ..ai.agents import (
-    PolicyAgent,
-    QAgent,
-    RandomAgent,
-    ImmediateWinBlockAgent,
-    FourThreePriorityAgent,
-    LongestChainAgent
-)
+from .pygame_utils import draw_board, CELL_SIZE
+from .agent_factory import create_agent
+from .result_utils import show_results
 
 # 学習済みモデルをまとめたフォルダ
 # 上位ディレクトリに配置されている ``models`` フォルダを取得
 MODEL_DIR = Path(__file__).resolve().parents[2] / "models"
-
-CELL_SIZE = 50                # 1マスあたりのピクセルサイズ
-LINE_COLOR = (0, 0, 0)        # グリッド線(黒)
-BG_COLOR = (222, 184, 135)    # 碁盤っぽい背景色 (ベージュ)
-BLACK_STONE_COLOR = (0, 0, 0) # 黒石
-WHITE_STONE_COLOR = (255, 255, 255) # 白石
-
-
-def draw_grid(screen, board_size):
-    """碁盤のグリッド線を描画するヘルパー"""
-    for i in range(board_size):
-        # 縦線の描画
-        start_pos = (i * CELL_SIZE + CELL_SIZE // 2, CELL_SIZE // 2)
-        end_pos = (
-            i * CELL_SIZE + CELL_SIZE // 2,
-            (board_size - 1) * CELL_SIZE + CELL_SIZE // 2,
-        )
-        pygame.draw.line(screen, LINE_COLOR, start_pos, end_pos, 2)
-
-        # 横線の描画
-        start_pos = (CELL_SIZE // 2, i * CELL_SIZE + CELL_SIZE // 2)
-        end_pos = (
-            (board_size - 1) * CELL_SIZE + CELL_SIZE // 2,
-            i * CELL_SIZE + CELL_SIZE // 2,
-        )
-        pygame.draw.line(screen, LINE_COLOR, start_pos, end_pos, 2)
-
-
-def draw_stones(screen, board):
-    """盤面上の石を全て描画する"""
-    board_size = board.shape[0]
-
-    # 石のある場所だけを抽出してループする
-    for x, y in np.argwhere(board != 0):
-        stone = board[x, y]
-        center_pos = (
-            y * CELL_SIZE + CELL_SIZE // 2,
-            x * CELL_SIZE + CELL_SIZE // 2,
-        )
-        color = BLACK_STONE_COLOR if stone == 1 else WHITE_STONE_COLOR
-        pygame.draw.circle(screen, color, center_pos, CELL_SIZE // 2 - 2)
-
-
-def draw_board(screen, env):
-    """背景・グリッド・石をまとめて描画するメイン関数"""
-    board_size = env.board_size
-    screen.fill(BG_COLOR)
-    draw_grid(screen, board_size)
-    draw_stones(screen, env.game.board)
-
 
 def play_game(env, black_agent, white_agent, visualize=True, fps=1):
     """
@@ -185,15 +129,14 @@ def play_agents_vs_agents(
         agent_path=white_agent_path,
         agent_params=white_agent_params
     )
-
-    # 複数対局を実行
+    # --- 複数対局を実行 -------------------------------------------------
     winners = []
     black_rewards = []
     turn_counts = []
 
     for game_idx in range(num_games):
         env = GomokuEnv(board_size=board_size, **env_params)
-        # 最初のゲームだけ表示する
+        # 最初のゲームのみ表示する
         do_visual = visualize and (game_idx == 0)
 
         winner, black_reward, turn_count = play_game(
@@ -201,13 +144,13 @@ def play_agents_vs_agents(
             black_agent=black_agent,
             white_agent=white_agent,
             visualize=do_visual,
-            fps=fps
+            fps=fps,
         )
         winners.append(winner)
         black_rewards.append(black_reward)
         turn_counts.append(turn_count)
 
-    # 集計結果を表示
+    # --- 結果表示 -------------------------------------------------------
     show_results(
         black_agent_type=black_agent_type,
         black_agent_path=black_agent_path,
@@ -215,99 +158,8 @@ def play_agents_vs_agents(
         white_agent_path=white_agent_path,
         winners=winners,
         black_rewards=black_rewards,
-        turn_counts=turn_counts
+        turn_counts=turn_counts,
     )
-
-# ------------------------------------------------------------------------------
-# エージェント生成用ヘルパー
-# ------------------------------------------------------------------------------
-def create_agent(agent_type, board_size, agent_path=None, agent_params=None):
-    """
-    agent_type の文字列に応じてエージェントを生成するヘルパー関数。
-
-    - "policy"/"q" などモデルを要するものはファイル読み込みにも対応。
-    - 該当しない文字列が渡された場合は RandomAgent を返す。
-    """
-
-    if agent_params is None:
-        agent_params = {}
-
-    # 文字列を小文字化しておく
-    key = agent_type.lower()
-
-    # 同義語をまとめる辞書
-    alias = {
-        "rand": "random",
-        "immediatewinblockagent": "immediate",
-        "fourthreepriorityagent": "fourthree",
-        "longestchainagent": "longest",
-    }
-    key = alias.get(key, key)
-
-    # エージェントクラスへのマッピング
-    class_table = {
-        "policy": PolicyAgent,
-        "q": QAgent,
-        "random": RandomAgent,
-        "immediate": ImmediateWinBlockAgent,
-        "fourthree": FourThreePriorityAgent,
-        "longest": LongestChainAgent,
-    }
-
-    if key in ("policy", "q"):
-        # 学習済みモデルを読み込む可能性があるエージェント
-        AgentClass = class_table[key]
-        agent = AgentClass(board_size=board_size, **agent_params)
-        if agent_path:
-            agent.load_model(agent_path)
-        return agent
-
-    if key in class_table:
-        # ヒューリスティック系 / RandomAgent
-        AgentClass = class_table[key]
-        return AgentClass()
-
-    # 万が一該当しなければ RandomAgent で代替
-    print(f"[WARNING] Unknown agent_type={agent_type}, fallback to RandomAgent.")
-    return RandomAgent()
-
-
-# ------------------------------------------------------------------------------
-# 結果表示用ヘルパー
-# ------------------------------------------------------------------------------
-def show_results(
-    black_agent_type,
-    black_agent_path,
-    white_agent_type,
-    white_agent_path,
-    winners,
-    black_rewards,
-    turn_counts
-):
-    """
-    複数試合の結果を集計してprint。
-    """
-    import numpy as np
-
-    num_games = len(winners)
-    avg_reward = float(np.mean(black_rewards))
-    black_win_rate = float(np.mean([1.0 if w == 1 else 0.0 for w in winners]))
-    avg_turn_count = float(np.mean(turn_counts))
-
-    print("============================================================")
-    print(f"対局数: {num_games}")
-    print(f"黒番 = {black_agent_type}, path={black_agent_path}")
-    print(f"白番 = {white_agent_type}, path={white_agent_path}")
-    print(f"黒番 平均報酬(視点) : {avg_reward:.3f}")
-    print(f"黒番 勝率          : {black_win_rate:.3f}")
-    print(f"平均決着手数       : {avg_turn_count:.1f}")
-    print("------------------------------------------------------------")
-
-    show_count = min(num_games, 10)
-    print(f"(先頭{show_count}ゲームの結果) => (winner, black_reward, turn)")
-    for i in range(show_count):
-        print(f"Game {i+1:2d}: Winner={winners[i]}, RewardForBlack={black_rewards[i]}, Turn={turn_counts[i]}")
-    print("============================================================")
 
 
 # ------------------------------------------------------------------------------
